requirements.txt
disnake
flask
asyncio
requests
python-dotenv
